tryCatch exceptions in solve(opt$hessian)
Use tryCatch to handle exceptions arising from calculating the variance-
covariance matrix. In particular, if the Hessian is not positive semi-
definite then there will be negative elements in the var-cov matrix.
In that case, issue a warning and set varcov to NA. Similarly, an
exactly singular Hessian causes as an error: set varcov elements to NA.
Output variance-covariance matrix from optLBSPRDome.